+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = "b12fdb91a5ca7ddc3585f5928fe768f50222f139"
+__version__ = "b93ab5ebcf05c9ee342e775cbc9b7efb64b18a8b"
